Add ability to override s3 endpoint URL for local development
Adds kwargs to dmutils.s3.S3, allowing consumers to pass arbitrary
arguments to the boto3 resource constructor.

If running Flask in development mode dmutils.s3.S3 will also
automatically use the url in `S3_ENDPOINT_URL` as the endpoint URL for
boto3.

This is useful for local testing, since setting `endpoint_url=` for the
resource allows using tools such as [localstack] when real AWS access is
not available or desirable.

[localstack]: https://github.com/localstack/localstack