Update date validator to highlight all fields if one field has an error
Design System guidance is that if more than one field for a date input
has an error then we should highlight all the fields.

This commit updates the tests and logic of DateValidator to meet this
requirement; I missed it earlier.

We put this logic in the `_error` helper method of DateValidator so that
it is one place and easier to change later.