Fix bugs in GreaterThan validator
While working on another ticket I found some bugs in the GreaterThan
validator:

- it will raise an exception if there is no data for field
- it will not compare itself to 0

This commit adds some tests and fixes the bugs.

Now the GreaterThan validator will not compare itself if has no input
(same as we were doing if other had no input), and will do the
comparison if other is 0.