Refactor Notify email logging
Logging messages about Notify emails is annoying, there are lots of
little variables that need to get passed around. This commit adds a new
class `DMNotifyEmail` to carry all the bits of information about an
email that might be relevant for logging, and creates a new method
`DMNotifyClient._log()` that logs a message about a DMNotifyEmail.

We then rewrite the existing logging statements and methods using
_log().

This change is backwards-compatible.