Add a message to be logged when an email is sent successfully
We get a lot of user support queries asking whether an email was sent
correctly or not; it would be nice to be able to always see a log
message saying "email was sent". This commit adds a log message to the
`DMNotifyClient.send_email_message()` method, so whatever logging there
may be around the call you can always be sure at least one message was
logged whether the email was sent successfully or not.