Rewrite DirectPlusClient.get_organization_by_duns_number()
We are having an issue where users can create supplier accounts with
invalid DUNS numbers. This appears to be because although we expect the
Direct+ error code `10001` for DUNS number not found, we are actually
getting `40105`.

This commit fixes this by changing the code to ignore the Direct+ error
code and just look at the HTTP status code to determine whether the
DUNS number is valid/found or not.

Also we now:

- log the HTTP status code along with the API response
- always log a message if the HTTP status code is an error code
- raise an exception instead of returning None (breaking change)
- raise a custom exception if there is an error instead of raising
  ValueError or KeyError