Add helper method to support assets rewrite performed by multiple apps
Supplier frontend has an authenticated endpoint for declaration documents: frameworks.download_declaration_document

Sometimes other apps need to be able to show declaration documents to users who are not the supplier who uploaded
the document. This function allows them to convert the supplier frontend URL into an assets URL.

You must ensure that you are only exposing this document to authorised users.

We're going to use this to give admins access. We can then update buyer frontend to use this (instead of hard-coding URLs)