Stop using dependabot for development dependencies
This matches how we use dependabot for the frontend apps. Currently, the bulk of dependabot PRs for this repo are for development dependencies (https://github.com/alphagov/digitalmarketplace-utils/pulls). I don't think keeping up to date with every patch version increment is a good use of our time and effort.