Hide DOS6 banner once it's closed
We think it's best to remove the banner in this period. Telling users something is coming soon and giving a link that they cannot use might cause confusion. Sometimes giving information when a user wouldn't normally be looking for it (out of context) can raise more questions than answers.

We'll need to rework this for G13 - I'm not doing that now because I don't know the equivalent dates for G13.