Log stack trace for failure to remove email from Mailchimp
Unlike all the other methods on the mailchimp client, this one is only called from scripts. This means that logging is unstructured, so we lose the information in `extra`.

Switch to `logger.exception` so that we get the stack trace, including the error details. This should make future errors easier to debug. Since emails are hashed, there's no risk of leaking PII.