Count removal as success if already removed
Now that we've got two environments (1.5 and 1.0), we're running the data retention script for both. The environments have different databases, but share Mailchimp lists. This means that we have the potential for race conditions - one environment's job can delete a user from a list before the other tries to.

When this happens, we get a MailChimpError with a status of 404. So count this particular circumstance as a success.